requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Pillow==9.4.0
PyJWT==1.7.1
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0